requirements.txt
flask
google-genai
aubio
python-dotenv